Add a common file
This moves ID, prefix, and toID from `app.ts` to `common.ts`,
because none of these types/function/variables should be in
an "app" file whose main purpose is to get the bot up and
running.

I vaguely recall a discussion in the Smogon Development Discord server
where global variables where decided against, so this seems like a
good compromise.