Move typescript & type decl to devDeps
`dependencies` should contain modules used in runtime; `devDependencies`
should contain modules used for developing the application.

PostgresSQL falls under the former, while transpilers (such as tsc)
and test frameworks (such as Jest) fall under the latter.

This can be seen in the official npm docs:
https://docs.npmjs.com/files/package.json#dependencies where it says:
> "Please do not put test harnesses or transpilers in your dependencies object."

In the `devDependencies` of the same docs it states:
>>> "If someone is planning on downloading and using your module in their program,
then they probably don’t want or need to download and build the external test or
documentation framework that you use.